move MD_Location to dedicated module
