+++ trlc/lexer_md.py
@@ -74,6 +74,7 @@ from fractions import Fraction
 
 from trlc.lexer import Token, TRLC_Lexer
 from trlc.errors import Location, Message_Handler
+from trlc.lexer_md_location import MD_Location
 
 
 class MD_Source_Reference(Location):
@@ -92,47 +93,6 @@ class MD_Source_Reference(Location):
         leading = len(self._source_line) - len(stripped)
         caret_col = max(col - 1 - leading, 0)
         return [stripped, " " * caret_col + "^"]
-
-
-class MD_Location(Location):
-    """Location for markdown-emitted STRING tokens.
-
-    Carries the raw string value and exposes ``text()`` so nested lexers can
-    treat markdown strings like regular TRLC string literals.
-    """
-
-    def __init__(
-        self,
-        file_name,
-        line_no,
-        col_no,
-        source_line="",
-        token_text="",
-        mh=None,
-    ):
-        super().__init__(file_name, line_no, col_no)
-        self._source_line = source_line
-        self._token_text = token_text
-        self._full_text = f'"""{self._token_text}"""'
-
-        # Nested_Lexer.source_location() expects these Source_Reference-like
-        # attributes to exist on string literal locations.
-        if mh is not None:
-            self.lexer = TRLC_Lexer(mh, file_name, self._full_text)
-            self.start_pos = 0
-            self.end_pos = len(self._full_text) - 1
-
-    def context_lines(self):
-        if not self._source_line:
-            return []
-        col = self.col_no if self.col_no else 1
-        stripped = self._source_line.lstrip()
-        leading = len(self._source_line) - len(stripped)
-        caret_col = max(col - 1 - leading, 0)
-        return [stripped, " " * caret_col + "^"]
-
-    def text(self):
-        return self._full_text
 
 
 class MD_Lexer(TRLC_Lexer):
